pyproject.toml
[tool.poetry]
name = "python-monorepo"
version = "0.1.0"
description = ""
authors = ["Teller Novel Inc. <dev@tellernovel.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
httpx = "^0.25.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pysen = "^0.10.5"
pytest = "^7.4.3"
pyyaml = "^6.0.1"
typer = "^0.9.0"
types-pyyaml = "^6.0.12.12"

[tool.poetry.group.train]
optional = true
[tool.poetry.group.train.dependencies]
jupyter = "^1.0.0"

[tool.poetry.group.proj_a.dependencies]
fastapi = "^0.104.0"
lib-hello = { path = "libs/lib_hello", develop = true }
uvicorn = "^0.23.2"

[tool.poetry.group.proj_b.dependencies]
flask = "^3.0.0"
lib-hello = { path = "libs/lib_hello", develop = true }

[tool.poetry.group.proj_c.dependencies]
pytorch-ignite = "^0.4.13"
typer = "^0.9.0"

[tool.pysen]
version = "0.10"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_isort = true
enable_flake8 = true
enable_mypy = true
mypy_preset = "strict"
py_version = "py311"
[[tool.pysen.lint.mypy_targets]]
paths = ["."]

[tool.pysen.lint.source]
includes = ["."]

[tool.black] # automatically generated by pysen
# pysen ignores and overwrites any modifications
line-length = 88
target-version = ["py311"]

[tool.isort] # automatically generated by pysen
# pysen ignores and overwrites any modifications
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
